make mmr_leaves and mmr_proof to Optional value
Update lib.rs

refator error (#12)

Update to substrate v0.9.40 (#14)

* update substrate dep beefy-merkle-tree to binary-merkle-tree

* Update Cargo.toml

* fix clippy

* update

* update

* Update MmrLeaf

* Update mmr.rs

* Update lib.rs

* update

* update verfiy function

* fix update_state_in_multiple_steps and  update_state_in_multiple_steps

* fix update_state_works

* Move new function to LightClient

* add check public key start with 0x

* Update integration_test.rs

---------


Update errors.rs

Use encoded authority set proof
